requirements.txt
fastapi
uvicorn
pandas
openpyxl
langchain
openai
python-dotenv
mysql-connector-python
